Extract optional weight.
- Extract optional weight from coommon.h to reduce dependency on this header.
- Add test.
